Code-review cleanup pass before upstream PR
Style + code-quality follow-ups across the PvP-Honor / Auto-Attack /
ZoneConflict patches in preparation for upstreaming. No behavioural change.

CharacterCombat.cs
  - Drop the redundant PvpLogger field; use the inherited Logger from GameObject.
  - Downgrade PvP-kill/assist/death/respawn-counter messages from Info to Debug.
    Active PvP zones would otherwise flood the server log with one INFO line
    per kill + per assist.
  - Remove the four unused ResurrectX/Y/Z/Yaw fields (dead code — never read or
    written anywhere in this branch).
  - Resolve the victim's zone + ZoneConflict only once in DoDie and pass both
    into AwardPvpHonor (previously looked up twice on every PvP kill).
  - Extract WarZoneHonorLoss as a named constant instead of the inline literal 10.
  - Move the DiedInPvpWarZone flag assignment out of AwardPvpHonor into the
    caller — keeps the honor function side-effect free w.r.t. the death-context
    flags that CSResurrectCharacterPacket reads.
  - Drop ?. on ZoneManager.GetConflicts() — it never returns null.

ZoneConflict.cs
  - Drop the stray "#pragma warning disable IDE0005" that was added by the
    previous patch; all four usings are actually referenced.
  - Move every operational log (SetTimerTask scheduling, CheckTimer fired,
    no-NextStateTime warning) down to Debug. Only the actual state transition
    in SetState stays at Info.
  - Reformat the SetState log to a single concise line.

CSResurrectCharacterPacket.cs
  - Replace the inline magic buff IDs 1128 / 2385 / 4424 with new
    BuffConstants entries: WeakenedBody, RespawnCooldown, WarZoneLeech.
  - Collapse the four duplicated buff-application blocks into a single
    ApplyRevivalDebuffs() + ApplyBuff() helper pair. The 5-minute cooldown
    duration is now a const RespawnCooldownDurationMs.
  - Drop ?. on ZoneManager.GetConflicts() — it never returns null.

BuffConstants.cs
  - Add WeakenedBody=1128, RespawnCooldown=2385, WarZoneLeech=4424 entries
    consumed by CSResurrectCharacterPacket.

Holdable.cs
  - GetAttackAnimId no longer allocates a uint[] + List<uint> on every call.
    Replaced with a stackalloc Span<uint>[3] — auto-attacks happen on a tick
    timer, so per-tick GC pressure adds up at scale.
  - Extract DefaultFistAnimId const (was a bare 2u literal).
  - Wrap the modulo in Math.Abs to make GetAttackAnimId safe against
    AutoAttackIndex overflow into negative range.

UseAutoAttackSkillTask.cs
  - Use `var` for the slot switch (project style).
  - Range fallback when neither weapon nor skill template defines a range is
    now 3f (matches the fist range elsewhere) instead of the ad-hoc 4f.

Skill.cs
  - Use `var` for slot + leftHand locals (project style).
  - Drop the int cast on FireAnimId (it's uint now).

SCSkillFiredPacket.cs
  - Change FireAnimId from int to uint to match the underlying anim ID type
    and avoid a needless cast at the call site.

+++ AAEmu.Game/Core/Packets/G2C/SCSkillFiredPacket.cs
@@ -19,10 +19,10 @@ public class SCSkillFiredPacket : GamePacket
 
     /// <summary>
     /// The fire animation ID sent to the client.
-    /// Default = skill template's FireAnim ID. Caller can override for weapon-based
+    /// Default = skill template's FireAnim ID. Callers can override for weapon-based
     /// auto-attack animation (Skill.GetWeaponAttackAnimId) or NPC anim cycling.
     /// </summary>
-    public int FireAnimId { get; set; }
+    public uint FireAnimId { get; set; }
 
     public SCSkillFiredPacket(uint id, ushort tl, SkillCaster caster, SkillCastTarget target, Skill skill, SkillObject skillObject) : base(SCOffsets.SCSkillFiredPacket, 1)
     {
@@ -32,7 +32,7 @@ public class SCSkillFiredPacket : GamePacket
         _target = target;
         _skill = skill;
         _skillObject = skillObject;
-        FireAnimId = (int)(skill.Template.FireAnim?.Id ?? 0);
+        FireAnimId = skill.Template.FireAnim?.Id ?? 0;
     }
 
     public override PacketStream Write(PacketStream stream)
